Exercises/Exercise-5/requirements.txt
psycopg2-binary==2.9.1
pandas==1.1.5